% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{covApp}
\alias{covApp}
\title{Run OriLyt Coverage Decay Analyzer Shiny application. Optionally, load in your parameters along with the app.}
\usage{
covApp(directory = "/input", metadata = NULL)
}
\arguments{
\item{directory, }{Used to pre-load the directory where your .tab files are stored}

\item{metadata, }{Used to pre-load sample metadata (must be formatted as csv with columns "sample" and "group")}
}
\value{
shiny.appobj
}
\description{
Run OriLyt Coverage Decay Analyzer Shiny application. Optionally, load in your parameters along with the app.
}
